Transition to Heroku server and Player class as model
This checkin moves a lot of the initial Player creation logic into the
Player class. The signup screen has been modified to have a single
Create An Account button. The Player is actually now created on the
Heroku test server. There is a significant amount of cleanup remaining
and the logic for the creation of the first Post still isn't in place
yet.