Split TradePost into multiple trade post classes
TradePost is now TradePost, NPCTradePost, PlayerTradePost, MyTradePost
and ForeignTradePost. Each derived class serves a very specific
purpose. When we tried to have all the TradePosts in a single class, it
became very messy with many member variables not being used depending
on which type the instance was. Now, it should be much clearer