Fix server refresh bug
When the client is refreshing data from the server, we used to fire off
a bunch of requests in parallel, the problem is that these requests get
responded to out of order, which caused problems with the refresh,
including locking the UI in a way that required a restart. Now all
calls are made serially, so this can no longer recur.