Fixed TradePost annotation view failure to refresh, specifically homebase
* the issue is because trade-posts only allow one
observer-annotation-view at one time (to prevent memory leaks);
however, this resulted in the Preview's tradepost annotation view
stealing the observer spot from the Map's tradepost annotation. This
can happen in a Go Home situation (but not limited to that) when the
Preview annotation view gets added after the Map annotation view. So,
this was causing the Map annotation view to never get updated when
these game events happen.
* trade-post redraw is not the issue here; it is working as expected.
The remove/re-add was basically causing the Map annotation-view to
steal back to the observer slot. So, the force-redraw method and all
code related to that has been removed.